tests/wpt/tests/css/css-scroll-anchoring/anchoring-with-bounds-clamping-div.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

#scroller {
  height: 500px;
  width: 200px;
  overflow: scroll;
}
#changer { height: 1500px; }
#anchor {
  width: 150px;
  height: 1000px;
  overflow: scroll;
}

</style>
<div id="scroller">
  <div id="changer"></div>
  <div id="anchor"></div>
</div>
<script>

// Test that scroll anchoring interacts correctly with scroll bounds clamping
// inside a scrollable <div> element.
//
// There should be no visible jump even if the content shrinks such that the
// new max scroll position is less than the previous scroll position.

test(() => {
  var scroller = document.querySelector("#scroller");
  scroller.scrollTop = 1600;
  document.querySelector("#changer").style.height = "0";
  assert_equals(scroller.scrollTop, 100);
}, "Anchoring combined with scroll bounds clamping in a <div>.");

</script>
