LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/anchoring-with-bounds-clamping-div-expected.txt


FAIL Anchoring combined with scroll bounds clamping in a <div>. assert_equals: expected 100 but got 515
